refactor: move <FormError /> to form directory
