+++ src/components/routes/users/edit/Form.tsx
@@ -20,7 +20,7 @@ import FormSubmitBtn from "@/components/Btns/Submit";
 import useLoading from "@/lib/hooks/useLoading";
 import editOwnProfileAction from "@/lib/actions/editOwnProfile";
 import { useEffect, useState } from "react";
-import FormError from "@/components/FormError";
+import FormError from "@/components/Form/Error";
 import { toast } from "sonner";
 import { useRouter } from "next/navigation";
